upload code about salloc

+++ src/interface/src/rpc.rs
@@ -32,6 +32,6 @@ pub struct MessageMeta {
 #[derive(Debug, Clone, Copy, Serialize, Deserialize)]
 pub struct MessageTemplateErased {
     pub meta: MessageMeta,
-    pub shm_addr: u64,
-    pub shm_addr_remote: u64
+    pub shm_addr: usize,
+    pub shm_addr_remote: usize
 }
